% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_get_attribute.R
\name{gff_get_attribute}
\alias{gff_get_attribute}
\title{Parse gff field}
\usage{
gff_get_attribute(gffColumn, field, attrSep = ";")
}
\arguments{
\item{gffColumn}{The column of the Gff file to parse, typically 'attributes'. No default.}

\item{field}{A string representing the field with the value you want
(formatted as \code{field = value;}). No default.}

\item{attrSep}{A string representing the field separator (\code{;} in the standard
gff format). Defaults to \code{;}.}
}
\value{
A vector containing the specified field values of the same size as the
number of features in the gff file.
}
\description{
This function allows you to parse the complex 'attributes' column of a gff file
and get only feature names or IDs, for example.
Adapted from function in the \href{https://bioconductor.org/packages/release/data/experiment/html/davidTiling.html}{davidTiling experimental package}.
Also posted on the Bioconductor support system \href{https://support.bioconductor.org/p/24657/}{here}.
}
\section{Details}{

Typical use of this function will include first using \code{\link{gff_read}},
also in this package, to load gff file and then parsing the attributes field.
See examples below.
}
\examples{
gff_read(s288C_annotation_R64_modified.gff)

gff_get_attribute(gff$attributes, 'Name')

gff_get_attribute(gff$attributes, 'ID')
}

