% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_read.R
\name{gff_read}
\alias{gff_read}
\title{Read gff file}
\usage{
gff_read(gffFile, nRows = -1)
}
\arguments{
\item{gffFile}{The path to the Gff file to read. No default.}

\item{nRows}{Optional maximum number of rows to read (the exact argument to \code{nRows}
of \code{read.table()}). Defaults to all rows.}
}
\value{
Gff as data frame with appropriate column names and column data set to the
correct class.
}
\description{
This function allows you to read a gff file. It uses \code{read.table()} to load the
gff as a data frame, adds the column names and sets the correct column data classes.
Adapted from function in the \href{https://bioconductor.org/packages/release/data/experiment/html/davidTiling.html}{davidTiling experimental package}.
Also posted on the Bioconductor support system \href{https://support.bioconductor.org/p/24657/}{here}.
}
\examples{
\dontrun{
gff_read(s288C_annotation_R64_modified.gff)
}
}

