% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwglabr.R
\docType{package}
\name{hwglabr}
\alias{hwglabr}
\alias{hwglabr-package}
\title{hwglabr: A compilation of functions used frequently by Hochwagen lab members.}
\description{
The hwglabr package provides several functions commonly used for data analysis
in the Howchwagen lab, with a focus on ChIP-seq data analysis.
}
\details{
Even though some plotting functions are included, these are rather brittle. They
can be used as a first pass plotting or as starter code to be adapted to the
users' needs.
}

