% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opening_act_batch_run.R
\name{opening_act_batch_run}
\alias{opening_act_batch_run}
\title{Batch run of \code{\link{opening_act}} function}
\usage{
opening_act_batch_run(inputDataFile, useReadr = TRUE, runMetaORF = TRUE,
  userInput = TRUE)
}
\arguments{
\item{inputDataFile}{String indicating path to file containing instructions.
The file must be in CSV format with the following columns (column names in the file must
be reproduced exactly as below):
\enumerate{
  \item \strong{path} Path to wiggle data.
  \item \strong{relevantGenotype} String indicating the relevant strain mutations, to
  be used as the argument to \code{\link{opening_act}} of the same name.
  \item \strong{chipTarget} String indicating the ChIP target protein, to be used as
  the argument to \code{\link{opening_act}} of the same name.
  \item \strong{sampleID} The sample ID, to be used as the argument to
  \code{\link{opening_act}} of the same name.
}
No default.}

\item{useReadr}{Boolean indicating whether to use the much faster \code{read_tsv()} from
Hadley Wickham's readr package instead of base R's \code{read.table} when reading the wiggle
data. Used as the argument to \code{\link{readall_tab}} of the same name.
Defaults to \code{TRUE}.}

\item{runMetaORF}{Boolean indicating whether to run the meta ORF analysis. This analysis
typically takes about 30 minutes to run, so it may be useful to exclude it. Used as the
argument to \code{\link{opening_act}} of the same name. Defaults to \code{TRUE}.}

\item{userInput}{Boolean indicating whether to ask user to check the format of the
provided labels. Defaults to \code{TRUE}.}
}
\value{
The output of \code{\link{opening_act}} for each of the data sets included in the
batch. Check the documentation of \code{\link{opening_act}} for more detail.
}
\description{
This function will take a CSV file containing the path to multiple wiggle data folders
and argument labels and run \code{\link{opening_act}} function for each data set in the
file. It allows you to run a batch of data sets autmatically.
}
\examples{
\dontrun{
opening_act_batch_run(inputDataFile="~/Desktop/inputs.csv")
opening_act_batch_run(inputDataFile="~/Desktop/inputs.csv", useReadr = FALSE,
                      runMetaORF = TRUE, userInput = FALSE)
}
}

