% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_chromosomes.R
\name{order_chromosomes}
\alias{order_chromosomes}
\title{Order chromosome column by chromosome length}
\usage{
order_chromosomes(inputData, chrColumn, decreasing = FALSE)
}
\arguments{
\item{inputData}{Data frame with chromosome column and 16 rows (for example the
output of \code{\link{chr_coverage}}).
No default.}

\item{chrColumn}{String name of chromosome column to order by. No default.}

\item{decreasing}{Logical indicating whether to order in decreasing order, from
longest to shortest chromosome. Defaults to \code{FALSE}.}
}
\value{
Input data frame with rows ordered by chromosome length.
}
\description{
Given a data frame with 16 rows and a column with chromosome number, this function
returns the data frame with rows ordered by chromosome length.
}
\examples{
\dontrun{
ordered_df <- order_chromosomes(chrCov_df, 'seqname')

ordered_df <- order_chromosomes(chrCov_df, 'chr', decreasing=TRUE)
}
}
