% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_arrow.R
\name{plot_gene_arrow}
\alias{plot_gene_arrow}
\title{Plot genes on an R wiggle plot}
\usage{
plot_gene_arrow(geneEnd, geneLength, orientation, minima = 400,
  color = "black", height = 0.2, yPos = -1)
}
\arguments{
\item{geneEnd}{A numeric position on the x-axis of the 3' end of the gene. No default.}

\item{geneLength}{The length of the gene. No default.}

\item{orientation}{A value of 1 for genes on the Watson strand or -1 for genes on the Crick strand.
No default.}

\item{minima}{A minimal gene length before the arrow starts turning inside out.
Dependent on the range of the x-axis. Default is 400 bp.}

\item{color}{Define the color of the gene. Default is black.}

\item{height}{To adjust to the height of the gene bar. Default is 0.2.}

\item{yPos}{To adjust the position of the gene on the y-axis. Default is -1.}
}
\description{
This function allows you to you to add arrows representing genes to an R wiggle plot. \cr
Written by Tovah Markowitz.
}
\examples{
plot_gene_arrow(Red1_end, Red1_end-Red1_start, 1)
plot_gene_arrow(672823, 672823-670340, 1)
}

