% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wiggle.R
\name{plot_wiggle}
\alias{plot_wiggle}
\title{Wiggle data line plot}
\usage{
plot_wiggle(wiggleDataA, wiggleDataB, chrA, chrB, genome, yMax,
  color = "grey50", protein, onScreen = TRUE)
}
\arguments{
\item{wiggleDataA}{A data frame of wiggle data with two columns: genome position and signal. No default.}

\item{wiggleDataB}{Optional data frame of wiggle data with two columns: genome position and signal. No default.}

\item{chrA}{A number representing the chromosome of 'wiggleDataA'. No default.}

\item{chrB}{Optional number representing the chromosome of 'wiggleDataB'.
Must be provided if 'wiggleDataB' is used. No default.}

\item{genome}{A string representing the genome used for mapping. No default.}

\item{yMax}{Optional number to be used as the max Y scale value in the plots.
Particularly useful to plot two chromosomes on the same Y scale. No default.}

\item{color}{Optional R color. Defaults to 'grey50'.}

\item{protein}{A string representing the ChIPped protein. No default.}

\item{onscreen}{Boolean indicating plots should be returned to the screen (onScreen = TRUE)
or written to .png files (onScreen = FALSE). Defaults to TRUE.}
}
\value{
One or two line plots, either on screen or as png files (in the working directory).
}
\description{
This function allows you to plot wiggle data for one or two selected chromosomes.
It takes as input the output of 'window_smooth' (R data frame with two columns:
genome position and smoothed signal).
}
\examples{
plot_wiggle(WT_chr3, WT_chr5, 3, 5, genome = 'SK1', yMax = 5, color = 'red', protein = 'Red1', onScreen = TRUE)

plot_wiggle(chr1, chr10, 1, 10, genome = 'S288C', yMax = 5, color = 'black', protein = 'Rec8-HA', onScreen = FALSE)
}

