% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readall_tab.R
\name{readall_tab}
\alias{readall_tab}
\title{Load wiggle data}
\usage{
readall_tab(fileLocation, useReadr = TRUE, progressBar = TRUE,
  localCopy = FALSE, asBedGraph = FALSE)
}
\arguments{
\item{fileLocation}{A string with the path to the folder containing the wiggle data. No default.}

\item{useReadr}{Boolean indicating whether to use the much faster \code{read_tsv()} from
Hadley Wickham's readr package instead of base R's \code{read.table}. Defaults to \code{TRUE}.}

\item{progressBar}{Boolean indicating whether to display a progress bar (using R package
pbapply). Defaults to \code{TRUE}.}

\item{localCopy}{Boolean indicating whether to create local copy of target files before reading
data. If \code{TRUE} a local folder is automatically created and deleted after use. Use this argument
to avoid reading files directly from shared locations (namely LabShare). Defaults to \code{FALSE}.}

\item{asBedGraph}{Boolean indicating whether to return the data in bedGraph-like format instead of
R list of 16 data frames. Genomic coordinates are kept as 1-based. Defaults to \code{FALSE}.}
}
\value{
Either an R list of 16 data frames, one for each chromosome (\code{asBedGraph=FALSE})
or a bedGraph-like data frame (\code{asBedGraph=TRUE}).
}
\description{
This function allows you to load tab-separated wiggle data.
Adapted from function written by Tovah Markowitz (original function name: \code{readAll.tab}).
}
\examples{
\dontrun{
readall_tab("/Path/to/wiggles/folder/Red1_WT_reps_SacCer3_2mis_MACS_wiggle_norm/")
readall_tab("/Path/to/wiggles/", useReadr = TRUE, progressBar = TRUE)
readall_tab("/Path/to/wiggles/", useReadr = T, progressBar = T, localCopy = T)
#' readall_tab("/Path/to/wiggles/", asBedgraph = TRUE)
}
}
