% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_centromeres.R
\name{remove_centromeres}
\alias{remove_centromeres}
\title{Remove centromeric regions from wiggle data}
\usage{
remove_centromeres(wiggleData, regionSize = 50000)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}). No default.}

\item{regionSize}{Number indicating the size (in bp) of the region to remove
(centered on the centromere of each chromosome). Defaults to 50'000 bp.}
}
\value{
A similar list of the 16 chr wiggle data after removal of the rows for
the centromeric regions (both positions and signal) in each chromosome data frame.
}
\description{
This function allows you to remove the region containing the centromere in each
chromosome. The original motivation was to analyse Red1 ChIP-seq data in rec8
mutants (show a strong enrichment around the centromeres).
}
\examples{
remove_centromeres(rec8)

remove_centromeres(rec8, regionSize = 30000)
}

