% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_conv.R
\name{signal_at_conv}
\alias{signal_at_conv}
\title{Signal between all convergent genes genomewide}
\usage{
signal_at_conv(inputData, region_size = 1000, saveFile = FALSE,
  input_dataFrame = FALSE)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of readall_tab) or a data frame (in
which case you must set 'input_dataFrame = TRUE'). No default.}

\item{region_size}{Number indicating the size (in bp) of the region to calculate.
Defaults to 1000 bp (+/- 500 bp).}

\item{saveFile}{Boolean indicating whether output should be written to a .txt file (in current working
directory). If 'saveFile = FALSE', output is returned to screen or an R object (if assigned).
Defaults to FALSE.}

\item{input_dataFrame}{Boolean indicating whether input data is a data frame. This is the case when you
have data loaded from a .bed format, typically nucleosome signal, as opposed to the standard wiggle data
in a list of 16 chromosomes. Defaults to FALSE.}
}
\value{
A local data frame (dplyr data frame) with three columns: chromosome, position (relative to midpoint
of intergenic region) and signal.
}
\description{
This function allows you to pull out the ChIP signal centered on midpoints of convergent genes.
It takes as input either the wiggle data as list of 16 chromosome (output of 'readall_tab()')
or complete genome in one data frame (for example loaded from .bed files).
}
\examples{
signal_at_conv(WT)

signal_at_conv(WT, region_size = 1500, saveFile = TRUE, input_dataFrame = FALSE)
}

