% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_conv_average_plot.R
\name{signal_at_conv_average_plot}
\alias{signal_at_conv_average_plot}
\title{Line plot of average signal between convergent genes}
\usage{
signal_at_conv_average_plot(inputDataA, inputDataB, genome, yMax,
  onScreen = TRUE, legendXcoord = xMin + xMin * 0.2, legendYcoord = yMax +
  yMax * 0.05, colorA = "grey50", colorB = "orange")
}
\arguments{
\item{inputDataA}{A data frame of average signal between convergent genes: relative
position and average signal. No default.}

\item{inputDataB}{Optional data in the same format for a second sample. No default.}

\item{genome}{A string representing the genome used for mapping. This is used in the title
of the plot only. No default.}

\item{yMax}{Optional number to be used as the max Y scale value in the plot.}

\item{onScreen}{Boolean indicating plots should be returned to the screen
(\code{onScreen = TRUE}) or written to .png files (\code{onScreen = FALSE}).
Defaults to \code{TRUE}.}

\item{legendXcoord}{A number representing the X coordinate to locate legend.
Defaults to minimum X (left-aligned).}

\item{legendYcoord}{A number representing the Y coordinate to locate legend.
Defaults to maximum Y (top-aligned).}

\item{colorA}{Optional R color for sample A. Defaults to \code{grey50}.}

\item{colorB}{Optional R color for sample B. Defaults to \code{green}.}
}
\value{
A line plot of one or two samples, either on screen or as a .png file
(in the working directory).
}
\description{
\strong{Deprecated! Use \code{\link{signal_at_intergen_average_plot()}} instead.} \cr
This function allows you to make a line plot of the ChIP signal over the average
intergenic region centered on midpoints of convergent genes. It takes as input a
data frame containing the average signal centered on the midpoints of convergent
gene regions.
}
\details{
To generate the input for this function starting from an R list of wiggle data
for the 16 chromosomes you should run:
1.  \code{signal_at_conv()} to pull out the signal at every convergent gene region.

2. \code{signal_average()} to calculate the average signal over all regions.
}
\examples{
signal_at_conv_average_plot(WT_conv_mean_signal, genome = 'S288C')

signal_at_conv_average_plot(WT_conv_mean_signal, dot1_conv_mean_signal, genome = 'SK1',
                            yMax = 3, onScreen = FALSE, legendXcoord = -500,
                            legendYcoord = 1, colorA = 'red', colorB = 'green')
}

