% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_intergen.R
\name{signal_at_intergen}
\alias{signal_at_intergen}
\title{Signal at intergenic regions between convergent, divergent or tandem genes genome-wide}
\usage{
signal_at_intergen(inputData, inputDataFrame = FALSE, orientation = "conv",
  regionSize = 1000, includeOverlapping = TRUE, saveFile = FALSE)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of \code{readall_tab}) or a data frame
(in which case you must set \code{inputDataFrame = TRUE}). No default.}

\item{inputDataFrame}{Boolean indicating whether input data is a data frame, as opposed to the
standard wiggle data in a list of 16 chromosomes. Defaults to \code{FALSE}.}

\item{orientation}{A string indicating the type of intergenic region to analyze, according to the
orientation of the flanking genes. Accepts one of:
\enumerate{
  \item \code{conv} Convergent genes
  \item \code{div} Divergent genes
  \item \code{tandem} Tandem genes. All regions in output will be in the same 5' -> 3' orientation
  (genes on Crick strand are reversed).
}
Defaults to \code{conv}.}

\item{regionSize}{Number indicating the size (in bp) of the region to calculate.
Defaults to 1000 bp (midpoint +/- 500 bp).}

\item{includeOverlapping}{Boolean indicating whether intergenic regions between overlapping genes should
be included in the analysis. If \code{includeOverlapping = FALSE}, intergenic regions between overlapping
genes will not be included in the analysis. Otherwise all regions are included. Defaults to \code{TRUE}.}

\item{saveFile}{Boolean indicating whether output should be written to a .txt file (in current working
directory). If \code{saveFile = FALSE}, output is returned to screen or an R object (if assigned).
Defaults to \code{FALSE}.}
}
\value{
A local data frame (dplyr data frame) with four columns:
\enumerate{
  \item \code{chr} Chromosome number
  \item \code{position} Nucleotide coordinate (relative to midpoint of intergenic region)
  \item \code{signal} ChIP signal
  \item \code{dist_apart} Distance between the two genes (negative value for overlapping genes)
}
}
\description{
This function allows you to pull out the ChIP signal over a window of positions of the selected size
(defaulting to \code{regionSize = 1000}) centered on midpoints of intergenic regions between genes of
the selected orientation: convergent, divergent or tandem.
It takes as input either the wiggle data as list of 16 chromosome (output of \code{readall_tab()})
or the complete genome in one data frame (for example loaded from .bed files).
}
\examples{
signal_at_intergen(WT, orientation = 'conv')

signal_at_intergen(WT, orientation = 'div', inputDataFrame = TRUE, regionSize = 1500,
                   includeOverlapping = FALSE, saveFile = TRUE)
}

