% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_orf.R
\name{signal_at_orf}
\alias{signal_at_orf}
\title{Signal at all ORFs genome-wide}
\usage{
signal_at_orf(inputData, gff, gffFile, saveFile = FALSE)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of \code{readall_tab()}). No default.}

\item{gff}{Optional dataframe of the gff providing the ORF cordinates. Must be provided if
\code{gffFile} is not. No default. Note: You can use the function \code{gff_read()} in hwglabr to
load your selected gff file.}

\item{gffFile}{Optional string indicating path to the gff file providing the ORF cordinates. Must be
provided if \code{gff} is not. No default.}

\item{saveFile}{Boolean indicating whether output should be written to a .txt file (in current working
directory). If \code{saveFile = FALSE'}, output is returned to screen or an R object (if assigned).
Defaults to FALSE.}
}
\value{
A local data frame with four columns: chr (chromosome number), position (normalized genome
between 0 and 2000 bp), signal (ChIP signal) and gene (systematic gene name).
}
\description{
This function allows you to pull out the ChIP signal over all ORFs in the genome. It collects the
signal over each ORF plus both flanking regions (1/2 the length of the ORF on each side) and
scales them all to the same value: 2 kb. This means that for two example genes with lengths of
500 bp and 2 kb, flanking regions of 250 bp and 1 kb, respectively, will be colleted up and
downstream. Both genes will then be scaled to 1 kb and all four flanking regions to 500 bp.
The function takes as input the wiggle data as a list of 16 chromosomes
(output of \code{readall_tab()}).
}
\examples{
signal_at_orf(WT, gff = gff)

signal_at_orf(WT, gffFile = S288C_annotation_modified.gff, saveFile = TRUE)
}

