% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_orf_average_plot.R
\name{signal_at_orf_average_plot}
\alias{signal_at_orf_average_plot}
\title{Line plot of average signal on ORFs}
\usage{
signal_at_orf_average_plot(inputDataA, inputDataB, genome, yMin, yMax,
  onScreen = TRUE, legendXcoord = -200, legendYcoord = yMax + yMax * 0.05,
  colorA = "grey50", colorB = "orange", smoothBandwidth = 0)
}
\arguments{
\item{inputDataA}{A data frame of average signal on ORFs +/- 1/2 the length of the ORF:
relative position and average signal. No default.}

\item{inputDataB}{Optional data in the same format for a second sample. No default.}

\item{genome}{A string representing the genome used for mapping. This is used in the title
of the plot only. No default.}

\item{yMin}{Optional number to be used as the minimum Y scale value in the plot.}

\item{yMax}{Optional number to be used as the maximum Y scale value in the plot.}

\item{onScreen}{Boolean indicating plots should be returned to the screen (onScreen = TRUE)
or written to .png files (onScreen = FALSE). Defaults to TRUE.}

\item{legendXcoord}{A number representing the X coordinate to locate legend.
Defaults to minimum X (left-aligned).}

\item{legendYcoord}{A number representing the Y coordinate to locate legend.
Defaults to maximum Y (top-aligned).}

\item{colorA}{Optional R color for sample A. Defaults to 'grey50'.}

\item{colorB}{Optional R color for sample B. Defaults to 'orange'.}

\item{smoothBandwidth}{Optional integer as the bandwith for smoothing data (for better-looking
line plots) using a Kernel Regression Smoother. Smoothing is performed using function 'ksmooth()'
from 'stats' package. For info on bandwith argument and the 'ksmooth()' function in general run
'?ksmooth'. Defaults to 0 (no smoothing).}
}
\value{
A line plot of one or two samples, either on screen or as a png file
(in the working directory).
}
\description{
This function allows you to make a line plot of the ChIP signal over the average
ORF. It takes as input a data frame containing the average signal over ORFs
(+/- flanking regions) genome-wide.
}
\details{
To generate the input for this function starting from an R list of wiggle data
for the 16 chromosomes you should run:

1. "signal_at_orf()" to pull out the signal at every ORF and flanking regions
(1/2 the length of the ORF).

2. "signal_average()" to calculate the average signal at each relative position
over all regions genome-wide.
}
\examples{
signal_at_orf_average_plot(WT_orf_S288C_mean_signal, genome = 'S288C')

signal_at_orf_average_plot(WT_orf_mean_signal, dot1_orf_mean_signal, genome = 'SK1',
                           yMax = 3, onScreen = FALSE, legendXcoord = -500,
                           legendYcoord = 1, colorA = 'red', colorB = 'green',
                           smoothBandwidth = 50)
}

