% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_summit.R
\name{signal_at_summit}
\alias{signal_at_summit}
\title{Signal around all bedfile midpoints}
\usage{
signal_at_summit(inputData, bedData, extension = 1000, onlyComplete = TRUE)
}
\arguments{
\item{inputData}{A list of 16 chr wiggle data (output of readall_tab). No default.}

\item{bedData}{A data frame of bed data to extract signal around. No default.}

\item{extension}{Number indicating the number of bases around the summit or midpoints.
Value will be added both upstream and downstream. Default is 1000 bp.}

\item{onlyComplete}{Many wiggle files are missing data at random positions across the
genome. Boolean indicates if you want to only keep regions where every base is 
included within the wiggle file. Default is \code{TRUE}.}
}
\value{
An R (dplyr) data frame with four columms: chromosome name, position (relative 
to midpoint), signal, and bed line number within its individual chromosome
}
\description{
This function allows you to pull out the ChIP signal centered around summits or 
midpoints of bedfiles if start and stop values are more than one base apart. It
takes as input wiggle data as a list of 16 chromosomes (output of\code{\link{readall_tab}})
and a bedfile determining the positions to extract. This function can also be used to
get signal around centromeres (see second example below). Use with \code{\link{signal_average}} 
to calculate the mean for every position around all midpoints. \cr
Written by Tovah Markowitz.
}
\examples{
\dontrun{
signal_at_summit(WT, red1_summit_bed)

# For signal around centromeres:
signal_at_summit(WT, S288Ccen, extension = 2e4, onlyComplete = FALSE)
}
}

