% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_from_telomeres.R
\name{signal_from_telomeres}
\alias{signal_from_telomeres}
\title{Collect signal from telomeres for all chromosome arms}
\usage{
signal_from_telomeres(inputData, lengthToCollect = 1e+05)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}). No default.}

\item{lengthToCollect}{Number specifying the length (in bp) of the region to collect signal for,
starting form the telomeres. Defaults to 100000 (i.e. 100 kb).}
}
\value{
A list with two elements, each one is itself a list containing the collected
signal separately for the right and left arms of the included chromosomes:
\enumerate{
  \item \code{small_chrs} List of data frames of collected signal for right and
  left arms of chromosomes 1, 3 and 6
  \item \code{large_chrs} List of data frames of collected signal for right and
  left arms of remaining chromosomes
}
Each data frame has two columns:
\enumerate{
  \item \code{distance_to_telomere} Distance to telomere in bp
  \item \code{signal} ChIP-seq signal
}
}
\description{
This function allows you to pull out the ChIP signal from all telomeres.\cr
The function takes as input the wiggle data as a list of 16 chromosomes.
(output of \code{\link{readall_tab}}). \cr\cr
\strong{Note:} Some chromosomes have higher positions in the wiggle data than the
length of the chromosome reported in the table, resulting in negative position values
appearing in the output data. This is a consequence of the extension transformation
performed by MACS and the values are not off by more than ~100 bp.\cr
\strong{Note:} The SK! genome annotation is currently not perfect and includes
several inconsistencies at telomeric and sub-telomeric regions. This must be kept in
mind when analying the data, as it will affect the output of this analysis.\cr
\cr \cr
}
\examples{
\dontrun{
signal_from_telomeres(WT)

signal_from_telomeres(WT, lengthToCollect = 50000)
}
}

