% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_from_telomeres_average.R
\name{signal_from_telomeres_average}
\alias{signal_from_telomeres_average}
\title{Compute mean signal and SD by position of output of \code{\link{signal_from_telomeres}}}
\usage{
signal_from_telomeres_average(inputData, separateSmallAndLarge = FALSE)
}
\arguments{
\item{inputData}{A list produced by calling \code{\link{signal_from_telomeres}}. No default.}

\item{separateSmallAndLarge}{Boolean specifying whether to generate a single data
frame for all chromosomes (\code{separateSmallAndLarge = FALSE}) or two data frames
separating small (chrs 1, 3, and 6) and large chromosomes (all remaining chromosomes).
Defaults to \code{FALSE}.}
}
\value{
If \code{separateSmallAndLarge = FALSE}, a dataframe with three columns:
\enumerate{
  \item \code{distance_to_telomere} Distance to telomere in bp
  \item \code{mean_signal} Average ChIP-seq signal
  \item \code{sd} Standard deviation
}
If \code{separateSmallAndLarge = TRUE}, a list of two dataframes (each with the
same three columns explained above):
\enumerate{
  \item \code{small_chrs} Data frame of chromosomes 1, 3 and 6
  \item \code{large_chrs} Data frame of remaining chromosomes
}
}
\description{
Given the  output of \code{\link{signal_from_telomeres}}, this function allows
you to compute the mean ChIP-seq signal and  standard deviation by position (as
distance to telomere) either for all chromosomes together or in two separate groups
for small (chrs 1, 3, and 6) and large chromosomes (all remaining chromosomes).\cr\cr
}
\examples{
\dontrun{
signal_from_telomeres_average(WT_signal)

signal_from_telomeres_average(signalFromTelomeres, separateSmallAndLarge = TRUE)
}
}

