% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_per_orf_length.R
\name{signal_per_orf_length}
\alias{signal_per_orf_length}
\title{Mean signal for all ORFs genome-wide}
\usage{
signal_per_orf_length(inputData, gff, gffFile, saveFile = FALSE)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}). No default.}

\item{gff}{Optional dataframe of the gff providing the ORF cordinates. Must be provided if
\code{gffFile} is not. No default. Note: You can use the function \code{\link{gff_read}} in hwglabr to
load your selected gff file.}

\item{gffFile}{Optional string indicating path to the gff file providing the ORF cordinates. Must be
provided if \code{gff} is not. No default.}

\item{saveFile}{Boolean indicating whether output should be written to a .txt file (in current working
directory). If \code{saveFile = FALSE}, output is returned to screen or an R object (if assigned).
Defaults to FALSE.}
}
\value{
A data frame equivalent to the supplied gff table with the following additional columns:
\enumerate{
  \item \code{length} Length of the ORF in bp
  \item \code{mean_signal} Mean of the signal collected from the supplied data
}
\strong{Note:} Skipped genes are included in the output with 'NA' for \code{mean_signal}.
}
\description{
This function allows you to calculate mean ChIP signal for each ORF in the genome. The function
goes through all included features in the supplied gff data and for each one it adds ORF length
(in bp) and the mean of the signal collected from the supplied ChIP-seq data. This can then be
used to analyse mean signal as a function of ORF length.\cr
The function takes as input the wiggle data as a list of 16 chromosomes (output of
\code{\link{readall_tab}}).
\cr \cr
\strong{Note:} Our wiggle data always contains gaps with missing chromosome coordinates
and ChIP-seq signal. The way this function deals with that is by skipping affected genes.
The number of skipped genes in each chromosome is printed to the console, as well as the
final count (and percentage) of skipped genes. \cr
}
\examples{
\dontrun{
signal_per_orf_length(WT, gff = gff)

signal_per_orf_length(WT, gffFile = S288C_annotation_modified.gff, saveFile = TRUE)
}
}
