% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_compress.R
\name{wiggle_compress}
\alias{wiggle_compress}
\title{Compress wiggle data}
\usage{
wiggle_compress(inputWiggles, window)
}
\arguments{
\item{inputWiggles}{Will accept two types of wiggle information:
\enumerate{
  \item A single wiggle region/chromosome
  \item Multiple wiggle files to be directly compared/compressed. These must be
listed as strings in the following format:
\code{c("file1[[chr]]", "file2[[chr]]", ...)}.} \cr}

\item{window}{Size of window to be averaged across. Suggested size for
chromosome-wide plots for presenting and publication: 200 bp. No default.}
}
\value{
A R (dplyr) data frame with averaged position and value for each
}
\description{
This function will compress wiggle data by averaging values across
non-overlapping windows. It will take any number of wiggle files, but can
only handle one chromosome or chromosomal region per function call. This
function assumes that all data is from the chromosome and ignores all 
chromosomal information. Note: calls function compress. \cr
Written by Tovah Markowitz.
}
\examples{
wiggle_compress( red1[[6]], 200 )

wiggle_compress( c( "red1[[6]]", "hop1[[6]]", "smc4[[6]]" ), 200 )
}

