% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_correlation.R
\name{wiggle_correlation}
\alias{wiggle_correlation}
\title{Correlation between two ChIP-seq experiments}
\usage{
wiggle_correlation(Set1, Set2, method = "pearson")
}
\arguments{
\item{Set1}{A list of 16 chr wiggle data (output of \code{\link{readall_tab}})}

\item{Set2}{A list of 16 chr wiggle data (output of \code{\link{readall_tab}})}

\item{method}{A character string indicating which correlation coefficient
is to be computed (\code{method} argument to R function \code{cor()}).
Accepts one of \code{pearson} (default), \code{kendall}, or \code{spearman}.}
}
\value{
A vector with 17 values with names indicating the chromosome associated
with each individual correlation value.
}
\description{
This function allows you to calculate correlations between wiggle files
for ChIP-seq, both on an individual chromosome basis and genome-wide. It
takes as an input two lists of 16 chromosomes (output of \code{\link{readall_tab}}).
}
\examples{
wiggle_correlation(Red1WT_rep1, Red1WT_rep2)

wiggle_correlation(Red1WT_rep1, Red1WT_rep2, method = "spearman")
}

