% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_scatter.R
\name{wiggle_scatter}
\alias{wiggle_scatter}
\title{To make a scatter plot comparison of two wiggle files}
\usage{
wiggle_scatter(wiggleData1, wiggleData2, window = 5000)
}
\arguments{
\item{wiggleData1}{A list of 16 chr wiggle data (output of \code{readall_tab}). No default.}

\item{wiggleData2}{A second set of wiggle data (output of \code{readall_tab}). No default.}

\item{window}{Parameter of \code{wiggle_scatter}. Size of window to be compressed.
Default is 5000 bp.}
}
\value{
Output is a large (dplyr) data frame with
compressed data from all chromosomes. See \code{\link{wiggle_compress}} for more 
information. \cr
}
\description{
This function is designed to work with function \code{\link{wiggle_scatter_plot}} to
allow for an overall generalized comparison of two data sets.
Great for comparing replicates, checking controls, etc.
The two functions are used as follows:
\enumerate{
  \item \code{wiggle_scatter} will extract the necessary information using
  \code{\link{wiggle_compress}} for all chromosomes
  \item \code{\link{wiggle_scatter_plot}} will make a simple .eps file of the results.
}
Written by Tovah Markowitz
}
\examples{
a <- wiggle_scatter(red1, hop1, 4000)

wiggle_scatter_plot(a)
}

