% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_scatter_plot.R
\name{wiggle_scatter_plot}
\alias{wiggle_scatter_plot}
\title{To make a scatter plot comparison of two wiggle files}
\usage{
wiggle_scatter_plot(scatter, method = "spearman")
}
\arguments{
\item{scatter}{Input should be output of \code{\link{wiggle_scatter}}.}

\item{method}{Method used to calculate correlation value on plot (same as for
\code{\link{cor}} function). Defaults to "spearman".}
}
\value{
Output is an .eps file with file name and axis
names determined by columns of input data frame. The function is written so that
the x-axis and the y-axis have the same range.
}
\description{
This function is designed to work with function \code{\link{wiggle_scatter}} to
allow for an overall generalized comparison of two data sets.
Great for comparing replicates, checking controls, etc.
The two functions are used as follows:
\enumerate{
  \item \code{\link{wiggle_scatter}} will extract the necessary information using
  \code{\link{wiggle_compress}} for all chromosomes
  \item \code{wiggle_scatter_plot} will make a simple .eps file of the results.
}
Written by Tovah Markowitz
}
\examples{
a <- wiggle_scatter(red1, hop1, 4000)

wiggle_scatter_plot(a)
}

