% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_signal_average.R
\name{wiggle_signal_average}
\alias{wiggle_signal_average}
\title{Calculate genome-wide signal average in wiggle data}
\usage{
wiggle_signal_average(inputData)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of
\code{\link{readall_tab}}). No default.}
}
\value{
A decimal value of type numeric.
}
\description{
Given the ChIP-seq wiggle data as a list of 16 chromosomes (output of
\code{\link{readall_tab}}), this function calculates the genome-wide signal average.
}
\examples{
\dontrun{
wiggle_signal_average(wt_wiggle)
}
}
