% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_smooth.R
\name{wiggle_smooth}
\alias{wiggle_smooth}
\title{Sliding window smooth of wiggle data}
\usage{
wiggle_smooth(wiggleData, chrNumber, bandwidth = 200, useKsmooth = FALSE)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of readall_tab). No default.}

\item{chrNumber}{A number representing the chromosome to smooth. No default.}

\item{bandwidth}{A number representing the length of the smoothing window in bp
(or the Gaussian kernel bandwith, if "useKsmooth = TRUE"). Defaults to 200.}
}
\value{
An R data frame with two columns: genome position and smoothed signal.
}
\description{
This function allows you to smooth wiggle data using a sliding window.
}
\examples{
wiggle_smooth(WT, 1, 200)

wiggle_smooth(WT, 16, 100)
}

