% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_smooth.R
\name{window_smooth}
\alias{window_smooth}
\title{Sliding window smooth}
\usage{
window_smooth(wiggleData, chrNumber, windowSize = 200)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of readall_tab). No default.}

\item{chrNumber}{A number representing the chromosome to smooth. No default.}

\item{windowSize}{A number representing the smoothing window size (number of bp). Defaults to 200.}
}
\value{
An R data frame with two columns: genome position and smoothed signal.
}
\description{
This function allows you to smooth data using a sliding window.
}
\examples{
window_smooth(WT, 1, 200)

window_smooth(WT, 16, 100)
}

