% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_signal_at_conv.R
\name{average_signal_at_conv}
\alias{average_signal_at_conv}
\title{Calculate average signal between convergent genes}
\usage{
average_signal_at_conv(inputData, saveFile = FALSE)
}
\arguments{
\item{inputData}{As a data frame (output of 'signal_at_conv()'). No default.}

\item{saveFile}{Boolean indicating whether output should be written to a .txt file (in current
working directory). If 'saveFile = FALSE', output is returned to screen or an R object (if assigned).
Defaults to FALSE.}
}
\value{
An R data frame with two columns: position (relative to midpoint of intergenic region) and
mean signal.
}
\description{
This function allows you to calculate the ChIP signal over the average intergenic region centered
on midpoints of convergent genes. It takes as input a data frame containing the genome-wide signal
centered on the midpoints of convergent gene regions (output of 'signal_at_conv()') and calculates
the average for each relative position (e.g. between midpoint of intergenic convergent gene regions
+ and - 500 bp).
}
\examples{
average_signal_at_conv(WT_conv)

average_signal_at_conv(WT_conv, saveFile = TRUE)
}

