% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_coverage.R
\name{chr_coverage}
\alias{chr_coverage}
\title{Average signal per chromosome}
\usage{
chr_coverage(wiggleData)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}). No default.}
}
\value{
A 16x2 R data frame with two columns: chromosome number and average signal.
}
\description{
This function allows you to calculate the average signal (or coverage) per chromosome.
}
\examples{
chr_coverage(WT)

chr_coverage(rec8)
}

