% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_coverage.R
\name{chr_coverage}
\alias{chr_coverage}
\title{Average signal per chromosome}
\usage{
chr_coverage(wiggleData, removeCen = FALSE, cenRegionSize = 50000)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}).
No default.}

\item{removeCen}{Boolean indicating whether to remove regions around centromeres
(using function \code{\link{remove_centromeres}}). Defaults to \code{FALSE}.}

\item{cenRegionSize}{Number indicating the size (in bp) of the region to remove
(centered on the centromere of each chromosome). Corresponds to argument \code{regionSize}
of \code{\link{remove_centromeres}}. Defaults to 50'000 bp.}
}
\value{
A 16x2 R data frame with two columns: chromosome number and average signal.
}
\description{
This function allows you to calculate the average signal (or coverage) per chromosome.
}
\examples{
\dontrun{
chr_coverage(WT)

chr_coverage(dot1, removeCen = TRUE, cenRegionSize = 50000)
}
}
