% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opening_act.R
\name{opening_act}
\alias{opening_act}
\title{Standard analysis of HT genomics experiments}
\usage{
opening_act(wiggleData, relevantGenotype, chipTarget, sampleID,
  userInput = TRUE, runMetaORF = TRUE)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}).
No default.}

\item{relevantGenotype}{String indicating the relevant strain mutations. Just use "WT",
for example, if there are no relevant mutations. No default.}

\item{chipTarget}{String indicating the ChIP target protein. No default.}

\item{sampleID}{String indicating the sample ID, including the ID used in the
analysis pipeline (with a date) and the read mapping conditions (see examples below).
The function asks the user to check that the provided "sampleID" matches the required
format before proceeding with the analysis. No default.}

\item{userInput}{Boolean indicating whether to ask user to check the format of the
\code{sampleID} argument. Defaults to \code{TRUE}.}

\item{runMetaORF}{Boolean indicating whether to run the meta ORF analysis. This analysis
typically takes about 30 minutes to run, so it may be useful to exclude it.
Defaults to \code{TRUE}.}
}
\value{
A new folder in ".../LabShare/HTGenomics/Opening_act/" containing output plots
(as .pdf files) of the following analysis:
\enumerate{
  \item \strong{Chromosome size bias}
  \item \strong{Signal at centromeres}
  \item \strong{Signal flanking rDNA}
  \item \strong{Signal at sub-telomeric regions} (data mapped to S288c reference genome only)
  \item \strong{Signal at DSB hotspots} (data mapped to S288c reference genome only)
  \item \strong{Signal at axis binding sites} (data mapped to S288c reference genome only)
  \item \strong{Signal at meta ORF}
}
}
\description{
This function will run the lab's standard analysis of HT genomics experiments for which
tab-separated wiggle data is generated. It will call different functions in the package
to produce several .pdf files of analysis plots, written to a new folder in
".../LabShare/HTGenomics/Opening_act/".
\cr \cr
\strong{Note:} When running on data aligned to the SK1 genome three of the plots are
not produced: signal at sub-telomeric regions, signal at DSB hotspots and signal at
axis binding sites. The reason for skipping the sub-telomeric signal analysis is the
fact that the SK1 genome annotation contains inconsistencies at sub-telomeric regions.
The reason for skipping the other two analysis is the fact that the reference data for
DSB hotspots and Red1 binding sites were not available for SK1 at the time of writing
this function. It is probably redundant to run this (long!) analysis for both genomes
anyway, and using data mapped to the S288c reference genome should be preferred.\cr
}
\examples{
\dontrun{
opening_act(wiggleData=WT, relevantGenotype="WT", chipTarget="Red1",
            sampleID="AH119C-040114-sacCer3-2mis")
opening_act(set1_wiggle_data, "set1", "Red1", "AH8584b-16032016-sacCer3-2mis")
opening_act(rec8, "rec8", "Red1", "AH8115b-24042015-SacCer3-2mis",
            userInput=FALSE, runMetaORF=FALSE)
}
}

