% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readall_tab.R
\name{readall_tab}
\alias{readall_tab}
\title{Load wiggle data}
\usage{
readall_tab(fileLocation)
}
\arguments{
\item{fileLocation}{The path to the folder containing the wiggle data. No default.}
}
\value{
An R list of 16 data frames, one for each chromosome.
}
\description{
This function allows you to load tab-separated wiggle data.
}
\examples{
readall_tab(/Path/to/wiggle/files/folder)
}

