% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_at_orf.R
\name{signal_at_orf}
\alias{signal_at_orf}
\title{Signal at all ORFs genome-wide (meta ORF)}
\usage{
signal_at_orf(inputData, gff, gffFile, loessSpan = 0.05, saveFile = FALSE)
}
\arguments{
\item{inputData}{As a list of the 16 chr wiggle data (output of \code{\link{readall_tab}}). No default.}

\item{gff}{Optional dataframe of the gff providing the ORF cordinates. Must be provided if
\code{gffFile} is not. No default. Note: You can use the function \code{\link{gff_read}} in hwglabr to
load your selected gff file.}

\item{gffFile}{Optional string indicating path to the gff file providing the ORF cordinates. Must be
provided if \code{gff} is not. No default.}

\item{loessSpan}{Number specifying \code{span} argument for \code{loess} function (the smoothing
parameter alpha). This controls the degree of smoothing of the signal.
Defaults to \code{0.05}.}

\item{saveFile}{Boolean indicating whether output should be written to a .txt file (in current working
directory). If \code{saveFile = FALSE}, output is returned to screen or an R object (if assigned).
Defaults to \code{FALSE}.}
}
\value{
A local data frame with four columns:
\enumerate{
  \item \code{chr} Chromosome number
  \item \code{position} Nucleotide coordinate (in normalized total length of 1 kb)
  \item \code{signal} ChIP-seq signal at each position (1 to 1000)
  \item \code{gene} Systematic gene name
}
}
\description{
This function allows you to pull out the ChIP signal over all ORFs in the genome. It collects the
signal over each ORF plus both flanking regions (1/2 the length of the ORF on each side) and
scales them all to the same value (1000). This means that for two example genes with lengths of
500 bp and 2 kb, flanking regions of 250 bp and 1 kb, respectively, will be collected up and
downstream. Both gene lengths will then be scaled to 500 and all four flanking regions to 250.
After scaling, a loess model of the signal is built and used to output predictions of the signal
at each position between 1 and 1000. \cr
The function takes as input the wiggle data as a list of 16 chromosomes.
(output of \code{\link{readall_tab}}).
\cr \cr
\strong{Note:} Our wiggle data always contains gaps with missing chromosome coordinates
and ChIP-seq signal. The way this function deals with that is by skipping affected genes.
The number of skipped genes in each chromosome is printed to the console, as well as the
final count (and percentage) of skipped genes. \cr
}
\examples{
\dontrun{
signal_at_orf(WT, gff = gff)

signal_at_orf(WT, gffFile = S288C_annotation_modified.gff,
              loessSpan = 0.1, saveFile = TRUE)
}
}

