% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_scatter.R
\name{wiggle_scatter}
\alias{wiggle_scatter}
\title{To make a scatter plot comparison of two wiggle files}
\usage{
wiggle_scatter(wiggleData1, wiggleData2, window = 5000)
}
\arguments{
\item{wiggleData1}{Parameter of \code{wiggle_scatter}. A list of 16 chr wiggle data 
(output of \code{readall_tab}). No default.}

\item{wiggleData2}{Parameter of \code{wiggle_scatter}. A second set of wiggle data 
(output of \code{readall_tab}). No default.}

\item{window}{Parameter of \code{wiggle_scatter}. Size of window to be compressed.
Default is 5000 bp.}

\item{scatter}{Parameter for \code{\link{wiggle_scatter_plot}}. Input should be output of 
\code{wiggle_scatter}.}
}
\value{
Output of \code{wiggle_scatter} is a large (dplyr) data frame with
compressed data from all chromosomes. See \code{\link{wiggle_compress}} for more 
information. \cr
Output of \code{\link{wiggle_scatter_plot}} is an .eps file with file name and axis
names determined by columns of input data frame. The function is written so that
the x-axis and the y-axis have the same range.
}
\description{
This function is designed to work with function \code{\link{wiggle_scatter_plot}} to
allow for an overall generalized comparison of two data sets.
Great for comparing replicates, checking controls, etc.
The two functions are used as follows:
\enumerate{
  \item \code{wiggle_scatter} will extract the necessary information using
  \code{\link{wiggle_compress}} for all chromosomes
  \item \code{\link{wiggle_scatter_plot}} will make a simple .eps file of the results.
}
Written by Tovah Markowitz
}
\examples{
a <- wiggle_scatter(red1, hop1, 4000)

wiggle_scatter_plot(a)
}

