% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_scatter_plot.R
\name{wiggle_scatter_plot}
\alias{wiggle_scatter_plot}
\title{To make a scatter plot comparison of two wiggle files}
\usage{
wiggle_scatter_plot(scatter)
}
\arguments{
\item{scatter}{Parameter for wiggle_scatter_plot. Input should be output of 
wiggle_scatter.}

\item{wiggleData1}{Parameter of wiggle_scatter. A list of 16 chr wiggle data 
(output of \code{readall_tab()}). No default.}

\item{wiggleData2}{Parameter of wiggle_scatter. A second set of wiggle data 
(output of \code{readall_tab()}. No default.}

\item{window}{Parameter of wiggle_scatter. Size of window to be compressed.
Default is 5000 bp.}
}
\value{
Output of wiggle_scatter is a large (dplyr) data frame with
compressed data from all chromosomes. See wiggle_compress for more 
information. \cr
Output of wiggle_scatter_plot is an .eps file with file name and axis names
determined by columns of input data frame. The function is written so that
the x-axis and the y-axis have the same range.
}
\description{
This function is designed to work with function \code{wiggle_scatter()} to
allow for an overall generalized comparison of two data sets.
Great for comparing replicates, checking controls, etc.
The two functions are used as follows:
\enumerate{
  \item \code{wiggle_scatter()} will extract the necessary information using
  \code{wiggle_compress()} for all chromosomes
  \item \code{wiggle_scatter_plot()} will make a simple .eps file of the results.}
Written by Tovah Markowitz
}
\examples{
a <- wiggle_scatter( red1, hop1, 4000 )
wiggle_scatter_plot( a )
}

