% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle_smooth.R
\name{wiggle_smooth}
\alias{wiggle_smooth}
\title{Sliding window smooth of wiggle data}
\usage{
wiggle_smooth(wiggleData, chrNumber, bandwidth = 200, useKsmooth = FALSE)
}
\arguments{
\item{wiggleData}{As a list of the 16 chr wiggle data (output of readall_tab). No default.}

\item{chrNumber}{A number representing the chromosome to smooth. No default.}

\item{bandwidth}{A number representing the length of the smoothing window in bp
(or the Gaussian kernel bandwith, if \code{useKsmooth = TRUE}). Defaults to 200.}

\item{useKsmooth}{Boolean indicating choice of smoothing function:
\enumerate{
  \item \code{useKsmooth = FALSE}: use a simple sliding window smoother. Smoothing is performed by
  sliding a window of the specified size (\code{bandwidth} argument) over all genomic positions in
  the data and replacing the position values by the middle position and the signal values by
  their mean.
  \item \code{useKsmooth = TRUE}: use a Gaussian Kernel Regression Smoother. Smoothing is performed
  using function \code{ksmooth()} from 'stats' package using the default \code{normal} kernel and the
  specified bandwith.
  }
  Defaults to \code{FALSE}.}
}
\value{
An R data frame with two columns: genome position and smoothed signal.
}
\description{
This function allows you to smooth wiggle data using a sliding window.
}
\examples{
wiggle_smooth(WT, 1, 200)

wiggle_smooth(WT, 16, 100)
}

